# [shell] minor performance improvements