--- a/src/Core/tests/Benchmarks/Program.cs
+++ b/src/Core/tests/Benchmarks/Program.cs
@@ -6,8 +6,7 @@ namespace Microsoft.Maui.Handlers.Benchmarks
 	{
 		static void Main(string[] args)
 		{
-			//BenchmarkRunner.Run<MauiServiceProviderBenchmarker>();
-			BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).RunAllJoined();
+			BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
 		}
 	}
 }