Reviewing `dotnet trace` output of the .NET Podcast app:

    54.94ms Microsoft.NetConf2021.Maui!Microsoft.NetConf2021.Maui.Pages.MobileShell.InitializeComponent
    ...
    10.30ms Microsoft.Maui.Controls!Microsoft.Maui.Controls.ShellUriHandler.GetNavigationRequest
     2.73ms System.Linq!Enumerable.ToArray
     1.26ms System.Linq!Enumerable.Skip
     1.24ms Microsoft.Maui.Controls!Microsoft.Maui.Controls.Routing.GetRouteKeys

Reviewing the code, it looks like we can make several low-risk changes:

1. `Routing.GetRouteKeys()` can return a `HashSet<string>` as various
   usage of `keys.Contains()` will be much faster.

2. Cache the result of `Routing.GetRouteKeys()`. The cache is cleared
   when routes change.

3. Create a `List<string>` instead of doing
   `RetrievePaths(localPath).Skip(1).ToArray()`. The list can be used
   throughout instead of the array.

4. Use `Array.Empty<object>()` in a couple places instead of `new
   object[0]`.

Afterwards, `dotnet trace` output looks a bit better:

    7.37ms Microsoft.Maui.Controls!Microsoft.Maui.Controls.ShellUriHandler.GetNavigationRequest

System.Linq calls are gone from the trace. This seems like it would
save ~3ms of startup for MAUI app using shell on a Pixel 5 device.

I also added a benchmark for Shell navigation:

Before:

| Method |     Mean |    Error |   StdDev |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------- |---------:|---------:|---------:|-------:|-------:|------:|----------:|
| --GoTo | 75.06 us | 1.063 us | 0.994 us | 7.4463 | 0.6104 |     - |  61.13 KB |
| ++GoTo | 73.21 us | 0.907 us | 0.849 us | 7.3242 | 0.6104 |     - |  60.82 KB |

It appears the changes are faster with less allocations.